requirements.txt
alembic==1.13.3
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
black==24.8.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
coverage==7.11.1
Deprecated==1.3.1
dnspython==2.8.0
email_validator==2.2.0
flake8==7.1.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.21.3
marshmallow-sqlalchemy==1.1.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.11.2
mypy_extensions==1.1.0
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
phonenumbers==8.13.48
pillow==10.4.0
platformdirs==4.5.0
pluggy==1.6.0
psycopg==3.2.1
psycopg-binary==3.2.1
pycodestyle==2.12.1
pycparser==2.23
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
rich==13.9.4
shortuuid==1.0.13
six==1.17.0
SQLAlchemy==2.0.36
types-python-dateutil==2.9.0.20240316
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
wrapt==2.0.1
